Break up client.py into multiple files
The command-line interface remains unchanged, as all commands and options were
preserved, but this makes the code a little more manageable, as the original
client.py had grown to 1,200+ lines of code.
