+++ tests/test_client.py
@@ -1,4 +1,3 @@
-
 """
 test_client
 ----------------------------------
@@ -11,9 +10,7 @@ from __future__ import absolute_import
 
 import pytest
 from click.testing import CliRunner
-import os
 import re
-from io import StringIO
 
 try:
     from unittest.mock import patch, call
@@ -26,7 +23,6 @@ import alpenhorn.client as cli
 import alpenhorn.acquisition as ac
 import alpenhorn.archive as ar
 import alpenhorn.storage as st
-import alpenhorn.util as util
 
 import test_import as ti
 
@@ -43,7 +39,8 @@ def fixtures(tmpdir):
 
 @pytest.fixture(autouse=True)
 def no_cli_init(monkeypatch):
-    monkeypatch.setattr(cli, '_init_config_db', lambda: None )
+    monkeypatch.setattr(cli, 'config_connect', lambda: None )
+    monkeypatch.setattr(cli.node, 'config_connect', lambda: None )
 
 
 def test_import_schema(fixtures):
@@ -210,171 +207,17 @@ def test_status(fixtures):
     assert re.match(r'z\s+0\s+0\.0\s+bar.example.com:None', output[3])
 
 
-def test_verify(fixtures):
-    """Test the output of the 'verify' command"""
-    tmpdir = fixtures['root']
-
-    runner = CliRunner()
-
-    # test for error when mounting a non-existent node
-    result = runner.invoke(cli.verify, ['foo'])
-    assert result.exit_code == 1
-    assert 'Storage node "foo" does not exist.' in result.output
-
-    # test for error when check when the node is not active
-    node = st.StorageNode.get(name='x')
-    node.active = False
-    node.save()
-
-    result = runner.invoke(cli.verify, ['x'])
-    assert result.exit_code == 1
-    assert 'Node "x" is not active.' in result.output
-
-    # test for error when check for ALPENHORN_NODE fails
-    node.active = True
-    node.root = str(tmpdir)
-    node.save()
-
-    result = runner.invoke(cli.verify, ['x'])
-    assert result.exit_code == 1
-    assert 'Node "x" does not match ALPENHORN_NODE: '.format(node.root) in result.output
-
-    # test for 'x' when it is mounted, but contains no files
-    tmpdir.join('ALPENHORN_NODE').write('x')
-
-    result = runner.invoke(cli.verify, ['x'])
-    assert result.exit_code == 0
-    assert re.match(r'.*\n=== Summary ===\n' +
-                    '  0 total files\n' +
-                    '  0 missing files\n' +
-                    '  0 corrupt files',
-                    result.output, re.DOTALL)
-
-    # now pretend node 'x' has a copy of 'fred'
-    file_copy = (ar.ArchiveFileCopy
-                 .select()
-                 .join(ac.ArchiveFile)
-                 .where(ac.ArchiveFile.name == 'fred')
-                 .get())
-    file_copy.has_file = 'Y'
-    file_copy.save()
-    result = runner.invoke(cli.verify, ['x'])
-    assert result.exit_code == 2
-    assert re.match(r'.*\n=== Missing files ===\n' +
-                    str(tmpdir.join(file_copy.file.acq.name, file_copy.file.name)),
-                    result.output, re.DOTALL)
-    assert re.match(r'.*\n=== Summary ===\n' +
-                    '  1 total files\n' +
-                    '  1 missing files\n' +
-                    '  0 corrupt files',
-                    result.output, re.DOTALL)
-
-    ## now add a known file ('fred')
-    tmpdir.join('x', 'fred').write('')
-    result = runner.invoke(cli.verify, ['--md5', 'x'])
-    assert result.exit_code == 1
-    assert re.match(r'.*\n=== Corrupt files ===\n' +
-                    '/.*/ROOT/x/fred\n'
-                    '.*\n=== Summary ===\n' +
-                    '  1 total files\n' +
-                    '  0 missing files\n' +
-                    '  1 corrupt files',
-                    result.output, re.DOTALL)
-
-
-def test_clean(fixtures):
-    """Test the 'clean' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.clean, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Clean up NODE by marking older files as potentially removable.' in help_result.output
-    assert 'Options:\n  -d, --days INTEGER     Clean files older than <days>.' in help_result.output
-
-    ## pretend 'fred' is 1 GB in size
-    f = ac.ArchiveFile.get(ac.ArchiveFile.name == 'fred')
-    f.size_b = 1073741824.0
-    f.save()
-
-    # By default a FileCopy is set to has_file='N'
-    file_copy = (ar.ArchiveFileCopy
-                 .select()
-                 .join(ac.ArchiveFile)
-                 .where(ac.ArchiveFile.name == 'fred')).get()
-    file_copy.has_file = 'Y'
-    file_copy.save()
-
-    tmpdir = fixtures['root']
-    tmpdir.chdir()
-    result = runner.invoke(cli.clean, args=['-f', 'x'])
-    assert result.exit_code == 0
-    assert re.match(r'.*\nMark 1 files \(1\.0 GB\) from "x" available for removal\.\n.*' +
-                    r'Marked 1 files available for removal.\n',
-                    result.output, re.DOTALL)
-
-    ## by default, the cleaned copy should be marked as 'maybe wanted'
-    file_copy = (ar.ArchiveFileCopy
-                 .select()
-                 .join(ac.ArchiveFile)
-                 .where(ac.ArchiveFile.name == 'fred')).get()
-    assert file_copy.wants_file == 'M'
-
-    ## if we clean with the '--now' option, the copy should be marked as 'not wanted'
-    file_copy.wants_file = 'Y'
-    file_copy.save()
-    result = runner.invoke(cli.clean, args=['-f', '--now', 'x'])
-    assert result.exit_code == 0
-    assert re.match(r'.*\nMark 1 files \(1\.0 GB\) from "x" available for removal\.\n.*' +
-                    r'Marked 1 files available for removal.\n',
-                    result.output, re.DOTALL)
-
-    file_copy = (ar.ArchiveFileCopy
-                 .select()
-                 .join(ac.ArchiveFile)
-                 .where(ac.ArchiveFile.name == 'fred')).get()
-    assert file_copy.wants_file == 'N'
-
-    ## if we clean with the '--cancel' option, all unwanted copies should again be marked wanted
-    result = runner.invoke(cli.clean, args=['-f', '--cancel', 'x'])
-    assert result.exit_code == 0
-    assert re.match(r'.*\nMark 1 files \(1\.0 GB\) from "x" for keeping\.\n.*' +
-                    r'Marked 1 files for keeping.\n',
-                    result.output, re.DOTALL)
-
-    file_copy = (ar.ArchiveFileCopy
-                 .select()
-                 .join(ac.ArchiveFile)
-                 .where(ac.ArchiveFile.name == 'fred')).get()
-    assert file_copy.wants_file == 'Y'
-
-    ## '--cancel' and '--now' are mutually exclusive options
-    result = runner.invoke(cli.clean, args=['--now', '--cancel', 'x'])
-    assert result.exit_code == 1
-    assert 'Options --cancel and --now are mutually exclusive.' in result.output
-
-    # using a non-existent node should be reported as an error
-    result = runner.invoke(cli.clean, args=['--force', '--cancel', 'y'])
-    assert result.exit_code == 1
-    assert 'Storage node "y" does not exist.' in result.output
-
-    # cleaning an archive node without the force flag or interactive
-    # confirmation should be an error
-    result = runner.invoke(cli.clean, args=['z'])
-    assert result.exit_code == 1
-    assert 'Cannot clean archive node "z" without forcing.' in result.output
-
-
 def test_active(fixtures):
     """Test the output of the 'active' command"""
     runner = CliRunner()
 
-    help_result = runner.invoke(cli.active, ['--help'])
+    help_result = runner.invoke(cli.cli, ['active', '--help'])
     assert help_result.exit_code == 0
     assert 'List the nodes active on this' in help_result.output
     assert 'Options:\n  -H, --host TEXT  use specified host' in help_result.output
 
     # there are no files yet on the 'foo' host (i.e., on storage node 'x')
-    result = runner.invoke(cli.active, args=['--host', 'foo.example.com'])
+    result = runner.invoke(cli.cli, args=['active', '--host', 'foo.example.com'])
     assert result.exit_code == 0
     output = result.output.splitlines()
     assert len(output) == 1
@@ -391,190 +234,11 @@ def test_active(fixtures):
     file_copy.file.save()
 
     # now `active` should report one file
-    result = runner.invoke(cli.active, args=['--host', 'foo.example.com'])
+    result = runner.invoke(cli.cli, args=['active', '--host', 'foo.example.com'])
     assert result.exit_code == 0
     output = result.output.splitlines()
     assert len(output) == 1
     assert re.match(r'^x\s+' + str(fixtures['root']) + '\s+1 files$', output[0])
-
-
-@patch('alpenhorn.client.get_e2label', return_value=None)
-@patch('os.path.realpath', return_value=None)
-@patch('os.mkdir')
-@patch('subprocess.check_call', return_value=0)
-@patch('subprocess.check_output', side_effect=[
-    '',
-    '/dev/foo: LABEL="CH-fake-12-34-56-78" UUID="bar" TYPE="ext4"',
-    '',
-])
-@patch('glob.glob', return_value=['/dev/disk/by-id/fake-12-34-56-78'])
-@patch('os.getuid', return_value=0)
-def test_format_transport(getuid_mock, glob_mock,
-                          check_output_mock, check_call_mock,
-                          mkdir_mock, realpath_mock, get_e2label_mock, fixtures):
-    """Test the 'mount_transport' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.format_transport, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Interactive routine for formatting a transport disc as a storage node' in help_result.output
-    assert 'Options:\n  --help  Show this message and exit.' in help_result.output
-
-    result = runner.invoke(cli.format_transport, args=['12-34-56-78'])
-    assert result.exit_code == 0
-    assert re.match(r'.*\nDisc is already formatted\.\n' +
-                    r'Labelling the disc as "CH-12-34-56-78"' +
-                    r'.*\nSuccessfully created storage node.\n' +
-                    r'Node created but not activated. Run alpenhorn mount_transport for that.',
-                    result.output, re.DOTALL)
-    assert check_output_mock.mock_calls == [
-        call(['blkid', '-p', '/dev/disk/by-id/fake-12-34-56-78']),
-        call(['blkid', '-p', '/dev/disk/by-id/fake-12-34-56-78-part1']),
-        call(['df'])
-    ]
-    assert check_call_mock.mock_calls == [
-        call(['/sbin/e2label', '/dev/disk/by-id/fake-12-34-56-78-part1', 'CH-12-34-56-78'])
-    ]
-    assert mkdir_mock.mock_calls == [call('/mnt/CH-12-34-56-78')]
-    node = st.StorageNode.get(name='CH-12-34-56-78')
-    assert node.group.name == 'transport'
-    assert node.root == '/mnt/CH-12-34-56-78'
-    assert node.storage_type == 'T'
-
-
-@patch('alpenhorn.util.alpenhorn_node_check')
-@patch('os.path.ismount')
-@patch('os.system')
-def test_mount_transport(mock, mock_ismount, mock_node_check, fixtures):
-    """Test the 'mount_transport' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.mount_transport, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Mount a transport disk into the system and then make it available' in help_result.output
-    assert 'Options:\n  --user TEXT     username to access this node' in help_result.output
-
-    mock_node_check.return_value = True
-    mock_ismount.return_value = False
-    result = runner.invoke(cli.mount_transport, args=['z'])
-    assert result.exit_code == 0
-    assert mock.mock_calls == [call('mount /mnt/z')]
-    assert re.match(r'Mounting disc at /mnt/z',
-                    result.output, re.DOTALL)
-
-    mock_ismount.return_value = True
-    result = runner.invoke(cli.mount_transport, args=['x'])
-    assert result.exit_code == 0
-    assert mock.mock_calls == [call('mount /mnt/z')]
-    assert re.match(r'x is already mounted in the filesystem. Proceeding to activate it.',
-                    result.output, re.DOTALL)
-
-
-@patch('os.system')
-def test_unmount_transport(mock, fixtures):
-    """Test the 'mount_transport' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.unmount_transport, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Unmount a transport disk from the system and then remove it from alpenhorn.' in help_result.output
-    assert 'Options:\n  --help  Show this message and exit.' in help_result.output
-
-    result = runner.invoke(cli.unmount_transport, args=['x'])
-    assert result.exit_code == 0
-    assert mock.mock_calls == [call('umount /mnt/x')]
-    assert re.match(r'Unmounting disc at /mnt/x',
-                    result.output, re.DOTALL)
-
-
-def test_activate(fixtures):
-    """Test the 'activate' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.activate, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Interactive routine for activating a storage node located at ROOT.' in help_result.output
-    assert 'Options:\n  --path TEXT      Root path for this node' in help_result.output
-
-    # test for error when mounting a non-existent node
-    result = runner.invoke(cli.activate, args=['nonexistent'])
-    assert result.exit_code == 1
-    output = result.output.splitlines()
-    assert len(output) == 1
-    assert 'Storage node "nonexistent" does not exist. I quit.' in output[0]
-
-    # test for error when trying to mount a node that's already mounted
-    result = runner.invoke(cli.activate, args=['x'])
-    assert result.exit_code == 0
-    output = result.output.splitlines()
-    assert len(output) == 1
-    assert 'Node "x" is already active.' in output[0]
-
-    # now pretend the node is inactive so we can try to activate it
-    node = st.StorageNode.get(name='x')
-    node.active = False
-    node.save()
-
-    # test for error when check for ALPENHORN_NODE fails
-    result = runner.invoke(cli.activate, args=['x'])
-    assert result.exit_code == 1
-    assert 'Node "x" does not match ALPENHORN_NODE' in result.output
-    assert not st.StorageNode.get(name='x').active
-
-    # test for success when check for ALPENHORN_NODE passes and the node is
-    # mounted
-    x_root = fixtures['root'].join('x')
-    x_root.join('ALPENHORN_NODE').write('x')
-    result = runner.invoke(cli.activate,
-                           args=['--path=' + str(x_root),
-                                 '--user=bozo',
-                                 '--address=foobar.example.com',
-                                 'x'])
-    assert result.exit_code == 0
-    output = result.output.splitlines()
-    assert len(output) == 2
-    assert re.match(r'^I will set the host to ".+"\.$', output[0])
-    assert 'Successfully activated "x".' == output[1]
-
-    node = st.StorageNode.get(name='x')
-    assert node.active
-    assert node.root == x_root
-    assert node.username == 'bozo'
-    assert node.address == 'foobar.example.com'
-    assert node.host == output[0].split('"')[1] == util.get_short_hostname()
-
-
-def test_deactivate(fixtures):
-    """Test the 'deactivate' command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.deactivate, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Deactivate a storage node with location or named ROOT_OR_NAME.' in help_result.output
-    assert 'Options:\n  --help  Show this message and exit.' in help_result.output
-
-    result = runner.invoke(cli.deactivate, args=['x'])
-    assert result.exit_code == 0
-    output = result.output.splitlines()
-    assert len(output) == 1
-    assert 'Node successfully deactivated.' in output[0]
-    node = st.StorageNode.get(name='x')
-    assert not node.active
-
-    # deactivate already deactivated node
-    result = runner.invoke(cli.deactivate,
-                           args=['x'])
-    assert result.exit_code == 0
-    output = result.output.splitlines()
-    assert len(output) == 1
-    assert 'There is no active node there any more' in output[0]
-
-    # deactivate an unknown node
-    result = runner.invoke(cli.deactivate,
-                           args=['y'])
-    assert result.exit_code == 1
-    output = result.output.splitlines()
-    assert 'That is neither a node name, nor a path on this host. I quit.' == output[0]
 
 
 def test_import_files(fixtures):
@@ -953,63 +617,3 @@ def test_import_files_with_limiting(fixtures):
     assert inspect.cleandoc(expected_output) in result.output
 
     assert acq_file.copies.join(st.StorageNode).where(st.StorageNode.name == 'x').count() == 0
-
-
-def test_create_group(fixtures):
-    """Test the create group command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.create_group, ['--help'])
-    assert help_result.exit_code == 0
-    assert 'Create a storage GROUP' in help_result.output
-
-    tmpdir = fixtures['root']
-    tmpdir.chdir()
-    result = runner.invoke(cli.create_group, args=['group_x'])
-
-    assert result.exit_code == 0
-    assert result.output == 'Added group "group_x" to database.\n'
-    this_group = st.StorageGroup.get(name='group_x')
-    assert this_group.name == "group_x"
-
-    # create an already existing node
-    result = runner.invoke(cli.create_group, args=['foo'])
-    assert result.exit_code == 1
-    assert result.output == 'Group name "foo" already exists! Try a different name!\n'
-
-
-def test_create_node(fixtures):
-    """Test the create node command"""
-    runner = CliRunner()
-
-    help_result = runner.invoke(cli.create_node, ['--help'])
-    assert help_result.exit_code == 0
-    assert "Create a storage NODE within storage GROUP with a ROOT directory on\n  HOSTNAME." in help_result.output
-
-    tmpdir = fixtures['root']
-    tmpdir.chdir()
-
-    result = runner.invoke(cli.create_node, args=['y', 'root', 'hostname', 'bar'])
-    assert result.exit_code == 0
-    assert result.output == 'Added node "y" belonging to group "bar" in the directory "root" at host "hostname" to database.\n'
-
-    node = st.StorageNode.get(name='y')
-
-    assert result.exit_code == 0
-    assert node.group.name == 'bar'
-    assert node.name == 'y'
-    assert node.root == 'root'
-    assert node.host == 'hostname'
-
-    result = runner.invoke(cli.create_node, args=['y', 'root', 'hostname', 'baba'])
-
-    assert result.exit_code == 1
-    assert result.output == 'Requested group "baba" does not exit in DB.\n'
-
-    result = runner.invoke(cli.create_node, args=['x', 'root', 'hostname', 'bar'])
-    assert result.exit_code == 1
-    assert result.output == 'Node name "x" already exists! Try a different name!\n'
-
-    result = runner.invoke(cli.create_node, args=['--storage_type=Z', 'z', 'root', 'hostname', 'bar'])
-    assert result.exit_code == 2  # Click usage error
-    assert 'Invalid value for "--storage_type": invalid choice: Z. (choose from A, T, F)' in result.output
